tflight/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
numpy==1.21.0
pillow==8.3.1
tflite-runtime==2.5.0
python-multipart==0.0.5